su
I wanted to put this in the layout.  But I haven't figure out how I can
combine my table with the current table in the layout. When I click each
row in the current table, it blocks the entire screen so the map is not
shown.  @_@

I am still debugging my code.  Bug 1: the arrow does not indicate the
correct location when the table is sorted. Bug2: Red colored text in the
popup is not shown in the I.E.